requirements.txt
pure-python-adb>=0.3.0.dev0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pytest>=7.4.0
ultralytics>=8.0.0
